storage: Move to spice/state
No reason for internal/storage to be at the top level.
It's used only for spice/state.
