+++ internal/spice/state/storage/mem.go
