+++ internal/spice/state/store.go
@@ -8,7 +8,7 @@ import (
 	"io"
 
 	"github.com/charmbracelet/log"
-	"go.abhg.dev/gs/internal/storage"
+	"go.abhg.dev/gs/internal/spice/state/storage"
 )
 
 // DB provides a key-value store that holds JSON values.
